I've got notes working now, at least the first lane of them. Notes now fall according to my float, bool[] system and land at the judgement bar within a few milliseconds of when they should be there (can easily be tuned once everything is implemented and gameplay can be properly tested). Thinking about making them stay spawned a little while longer but turning the object view off so late note presses can still count for something, though every time I've tried to mess with object visibility it's been a bad time so I'm unsure.
It turns out what was wrong with my note spawning was that I was using Vector3s relative to the canvas object that my notes are children of, however the interpolation I'm doing only takes world position into account. Therefore, my notes were spawning where they were supposed to relative to the canvas and then teleporting way out to the left of the screen on the next frame, everything was working correctly though. I've set a rough position for the first note using an empty gameobject as a position reference to find the values I need in the editor.

Next up is fixing my other three note scripts and prefabs and then handling button input and collision checking.