Updated readME a bit and tried a new build and run of the game.
notes from the build and run: (1) note objects appear and do in fact move down (2) it seems like changing the window size or scaling keeps all UI elements functional and in place but moves the notes and screws up their interpolation.

this means that I need to find some sort of anchor to base my note movement off of potentially, or just figure out what the values need to be for the full resolution build of the game ( / figure out how and why they change).

I'm glad I'm getting through a lot of this troublesome engine construction and tuning very early on in this project. Hoping to have the entire last week before the presentation deadline to polish, meaning basic functionality has to be complete in like one to two weeks.