I added what I'm calling the "mania multiplier" / "mania system". If the player combo exceeds 100, a x2 multiplier is applied to score until combo is dropped or hits 250. When combo hits 250, the multiplier becomes x4 at that point and will continue until the player misses a note. It also adds a visual multiplier to the Score field in game (ie:   "Score: 1750" becomes "Score: 1750 x2", a bit ambiguous but I imagine that most players who have played a rhythm game before will understand that it is applying the multiple to current score additions rather than the total score. I plan to fine tune this once more songs are added and potentially add other possible multipliers.
As of right now I can't quite test this functionality without hard coding a value for combo > 100 or loosening the grading, as the best combo I've achieved is ~78. I am getting better at it though so I imagine that I will be able to get a combo of over 100 by the time the song is finished being charted.

An idea I had that I wanted to note here was a difficulty mode that doesn't change the amount of notes in the chart but instead makes the grading very strict. I plan to call this "Black Label" mode as another stylistic nod to bullet hell games. Black label generally means a stricter more difficult version of a game which (by the nature of the bullet hell genre) usually lends itself to finer, tighter movements in gameplay as the available screen space to dodge in is very low. I thought this was a great translation of this concept. I think some DJ Max games have a stricter judgement option but its a somewhat rare feature so I'm excited to implement it. I think that it wouldn't be entirely unrealistic to attempt to add this mode by the time I have to present this project.