Fixed bug in Conductor.cs were accuracy, score, and some other internal variables failed to reset after playing a song, causing inaccurate scoring on subsequent songs played.
Also, csvToMrc has reached v0.2. I reorganized the core functionality and added a menu loop that allows the user to convert multiple charts in a row without closing the program or reread and rewrite the last file to the specified location. This second part is especially helpful as it removes the need to restart the program and retype the file paths of the *.csv and *.txt files when working on a chart; i.e. I can set the read file and output destination and then continue to rewrite changes in the *.csv to a *.txt in Assets\Charts as I'm charting.